refactor(dashboard): split provider-capability-matrix monolith into 7 modules
968 LOC single-file component → directory with index.ts barrel +
data.ts (static data + helpers) + 5 sub-components (feature-matrix,
bfcl-rankings, wiring-gaps, anti-patterns, oas-provider-table).

Verified: tsc --noEmit clean, ESLint clean, Vite build success (13.97s).
